beach-accommodation-finder/requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.1